removed several ellip entries
